feat: add preconnect hint for Google Fonts stylesheet
This commit adds a <link rel="preconnect" /> resource
hint for the origin of the font stylesheet if the
preconnect option is enabled. This is necessary
because the font stylesheet and the font itself are
served from two separate origins and both should be pre-
connected.

See https://web.dev/font-best-practices/#preconnect-to-critical-third-party-origins for more information.